Coverity polish: 1605322 Overflowed integer argument
we do want all 32bits of seed to be passed to srand48(), and if the cast results
in a a negative number, that's fine. Let's see if Coverity can be appeased this
way, otherwise we'll probably need to ignore the report.

+++ lib/libvarnish/vrnd.c
@@ -196,5 +196,5 @@ VRND_SeedAll(void)
 	AZ(VRND_RandomCrypto(&seed, sizeof seed));
 	VRND_SeedTestable(seed);
 	AZ(VRND_RandomCrypto(&seed, sizeof seed));
-	srand48(seed);
+	srand48((long)seed);
 }
